+++ src/hooks/stamp/useDeleteStampChart.ts
