+++ src/hooks/stamp/useGetStampChart.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import { queryKeys } from "src/api/queryKeys";
 import { stampFetcher } from "src/api/stamp/fetcher";
-import type { IStampChartResponse } from "src/interfaces/challenges";
+import type { IStampChartResponse } from "src/interfaces/stamp";
 
 interface IUseGetChallengesProps {
   challengeId: number;
