+++ src/hooks/stamp/usePutStampImage.ts
