+++ src/pages/challenge-stamp/index.tsx
@@ -2,11 +2,11 @@ import { ReactNode, useEffect, useRef, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { ChevronRightIcon } from "src/assets/icons/ChevronRightIcon";
 import RoundButton from "src/components/roundButton";
-import { useGetInvitationCode } from "src/hooks/challenge/useGetInvitationCode";
 import { useGetPresignedUrl } from "src/hooks/challenge/useGetPresignedUrl";
-import { useGetRevivalTicketCount } from "src/hooks/challenge/useGetRevivalTicketCount";
-import { useGetStampChart } from "src/hooks/challenge/useGetStampChart";
-import { usePutStampImage } from "src/hooks/challenge/usePutStampImage";
+import { useGetInvitationCode } from "src/hooks/invitation/useGetInvitationCode";
+import { useGetRevivalTicketCount } from "src/hooks/stamp/useGetRevivalTicketCount";
+import { useGetStampChart } from "src/hooks/stamp/useGetStampChart";
+import { usePutStampImage } from "src/hooks/stamp/usePutStampImage";
 import useModal from "src/hooks/useModal";
 import theme from "src/lib/theme";
 import { CertificateImageUploadDialog, ForgiveDialog, InviteCodeCopyDialog, RevivalSuccessDialog, RevivalTicketDialog } from "./dialog";
